New NULL_CANDLE

+++ crates/models/src/constants.rs
@@ -58,7 +58,7 @@ pub mod candles {
         ],
     };
 
-    pub const NULL_CANDLE: u32 = 0xFFFF;
+    pub const NULL_CANDLE: u16 = 0xFFFF;
 }
 
 pub mod price_helper {
